+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Deprecate `p[metric|log|trace]otlp.RegiserServer` in favor of `p[metric|log|trace]otlp.RegiserGRPCServer` (#6180)
 
+### 💡 Enhancements 💡
+
+- Add config marshaler (#5566)
+
 ## v0.61.0 Beta
 
 ### 🛑 Breaking changes 🛑